+++ BAAAM/BAAAM/RandomSpeedObj.h
@@ -16,8 +16,8 @@ public:
 private:
 
 	// 속력
-	float m_SpeedX;
-	float m_SpeedY;
+	int m_SpeedX;
+	int m_SpeedY;
 
 	// 이동방향 구분
 	bool m_IsRight;
